+++ docs/changes.md
@@ -2,6 +2,7 @@
 
 2025/11/3
   * 优化debug(感谢兆坤提供补丁)
+  * 优化widget_set_self_layout_params，值发生变化时才触发layout(感谢雨欣提供补丁)
 
 2025/11/2
   * 修复潜在的内存问题。